// semFront/src/app/models/rehabilitacion-interface.ts
export class RehabilitacionInterface {
   id?: string;
   nombre?: string;
   apellido_paterno?: string;
   apellido_materno?: string;
   fecha_nacimiento?: string;
   sexo?: string;
   img_subrogacion?: string;
   img_vigencia?: string;
   img_orden?: string;
   img_identificacion?: string;
   img_citas?: string;
   img_acta?: string;
   img_curp?: string;
   img_ident_tutor?: string;
   img_bitacora?: string;
}